requirements.txt
# Equity Infra MVP
streamlit>=1.28.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
simfin>=0.3.0
